Dev tanmay

+++ database/migrations/2024_12_14_045303_create_comments_table.php
@@ -13,16 +13,18 @@ class CreateCommentsTable extends Migration
      */
     public function up()
     {
-        Schema::create('comments', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('blog_id')->nullable();
-            $table->unsignedBigInteger('user_id')->nullable();
-            $table->text('description')->nullable();
-            $table->timestamps();
-            // Adding foreign keys
-            $table->foreign('blog_id')->references('blog_id')->on('blog')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('comments')) {
+            Schema::create('comments', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('blog_id')->nullable();
+                $table->unsignedBigInteger('user_id')->nullable();
+                $table->text('description')->nullable();
+                $table->timestamps();
+                // Adding foreign keys
+                $table->foreign('blog_id')->references('blog_id')->on('blog')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -32,6 +34,8 @@ class CreateCommentsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('comments');
+        if (Schema::hasTable('comments')) {
+            Schema::dropIfExists('comments');
+        }
     }
 }
